+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### GHOSTERY 8.5.4 ()
+
++ Adds support for Ghostery Browser (#621, #622)
++ Sunset Ghostery Rewards (#630)
++ Updated feature names (#613, #626, #627)
++ Shortened metrics uninstall URL (#615)
++ Fixes bug with initial account settings sync (#629)
++ Bug fixes (#617, #624, #625, #628)
++ Updated translations
+
 ### GHOSTERY 8.5.3 (October 8, 2020)
 
 + Updated Firefox Android extension panel UI and mobile optimizations (#587)